requirements.txt
pandas
fastapi
uvicorn[standard]
python-multipart
google-generativeai
python-dotenv
sqlparse
pydantic
